fix: fix bug where launch tries to create an empty kubeconfig
