+++ internal/launch/constants.go
@@ -1,0 +1,17 @@
+/*
+Copyright (C) 2021-2023, Kubefirst
+
+This program is licensed under MIT.
+See the LICENSE file for more details.
+*/
+package launch
+
+const (
+	consoleURL        = "https://console.kubefirst.dev"
+	helmChartName     = "kubefirst"
+	helmChartRepoName = "kubefirst"
+	helmChartRepoURL  = "https://charts.kubefirst.com"
+	helmChartVersion  = "2.1.7"
+	namespace         = "kubefirst"
+	secretName        = "kubefirst-initial-secrets"
+)
